+++ src/main/java/kr/hhplus/be/commerce/infrastructure/config/order/OrderConfig.java
@@ -4,6 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import kr.hhplus.be.commerce.application.order.OrderPlaceProcessor;
+import kr.hhplus.be.commerce.application.order.OrderPlaceV2Processor;
 import kr.hhplus.be.commerce.domain.cash.repository.CashHistoryRepository;
 import kr.hhplus.be.commerce.domain.cash.repository.CashRepository;
 import kr.hhplus.be.commerce.domain.coupon.repository.UserCouponRepository;
@@ -40,7 +41,7 @@ public class OrderConfig {
 		MessageRepository messageRepository,
 		UserRepository userRepository
 	) {
-		return new OrderPlaceProcessor(
+		return new OrderPlaceV2Processor(
 			orderRepository,
 			paymentRepository,
 			productRepository,
